move APIFnQuery to action_query.go
